backend/requirements.txt
aniso8601==1.2.1
appdirs==1.4.3
astroid==1.5.3
autopep8==1.3.2
Babel==2.4.0
backports.functools-lru-cache==1.4
blinker==1.4
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
configparser==3.5.0
coverage==4.3.4
decorator==4.0.11
enum34==1.1.6
Flask==0.12.1
Flask-Admin==1.5.0
Flask-Babel==0.11.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-RESTful==0.3.5
Flask-SQLAlchemy==2.2
Flask-Uploads==0.2.1
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.2
flipflop==1.0
ForgeryPy==0.1
guess-language==0.2
gunicorn==19.7.1
hypothesis==3.8.3
idna==2.6
isort==4.2.15
itsdangerous==0.24
Jinja2==2.9.6
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
olefile==0.44
packaging==16.8
passlib==1.7.1
pbr==3.0.0
Pillow==4.3.0
pycodestyle==2.3.1
pylint==1.7.2
pyparsing==2.2.0
python-dateutil==2.6.0
python-openid==2.2.5
python-resize-image==1.1.11
pytz==2017.2
requests==2.18.4
singledispatch==3.4.0.3
six==1.10.0
SQLAlchemy==1.1.9
sqlalchemy-migrate==0.11.0
sqlparse==0.2.3
Tempita==0.5.2
urllib3==1.22
Werkzeug==0.12.1
Whoosh==2.7.4
wrapt==1.10.11
WTForms==2.1
